Adding Comments and GridSplitter Data class and enums
